fix: resolve issue with student header
