+++ app/(dashboard)/profile/page.tsx
@@ -4,8 +4,7 @@ import { getUserDetails } from "@/lib/actions/user.action";
 
 const page = async () => {
   const session = await auth()
-
-    const profileData = await getUserDetails(session?.user?.id!);
+  const profileData = await getUserDetails(session?.user?.id!);
 
   return (
     <ProfileCard personDetails={profileData}/>
